tests/mvp1/performance/requirements.txt
paho-mqtt>=2.0
confluent-kafka>=2.3
psycopg2-binary>=2.9
redis>=5.0
